// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const ReconciliationStatus = {
  PENDING: 'PENDING',
  MATCHED: 'MATCHED',
  DISCREPANCY: 'DISCREPANCY'
} as const

export type ReconciliationStatus = (typeof ReconciliationStatus)[keyof typeof ReconciliationStatus]


export const SyncProvider = {
  MSC: 'MSC',
  QBO: 'QBO',
  BLUEPRINT: 'BLUEPRINT'
} as const

export type SyncProvider = (typeof SyncProvider)[keyof typeof SyncProvider]


export const SyncStatus = {
  PENDING: 'PENDING',
  SUCCESS: 'SUCCESS',
  FAILURE: 'FAILURE'
} as const

export type SyncStatus = (typeof SyncStatus)[keyof typeof SyncStatus]